# Ingest metrics with Sentry as well