--- a/crates/symbolicator-service/Cargo.toml
+++ b/crates/symbolicator-service/Cargo.toml
@@ -29,7 +29,7 @@ moka = { version = "0.12.1", features = ["future", "sync"] }
 once_cell = "1.17.1"
 rand = "0.8.5"
 reqwest = { version = "0.11.0", features = ["gzip", "brotli", "deflate", "json", "stream", "trust-dns"] }
-sentry = { version = "0.32.1", features = ["tracing"] }
+sentry = { version = "0.32.1", features = ["tracing", "UNSTABLE_metrics"] }
 serde = { version = "1.0.137", features = ["derive", "rc"] }
 serde_json = "1.0.81"
 serde_yaml = "0.9.14"