The way we often `&X.to_string()` for `cadence` sake because it takes a `&'a str`, and now we have to `.to_string()` a shitton more because Sentry wants a `Cow<'static, str>` makes me die inside :-(

The overhead of using Sentry metrics looks like a ~2x-10x reduction in throughput:

```
# before
Workload 0 (concurrency: 10): 3140 operations, 209.33 ops/s
Workload 1 (concurrency: 10): 2496 operations, 166.40 ops/s
Workload 2 (concurrency: 10): 4581 operations, 305.40 ops/s
Workload 3 (concurrency: 100): 12127 operations, 808.47 ops/s
Workload 4 (concurrency: 100): 16652 operations, 1110.13 ops/s
Workload 5 (concurrency: 50): 17048 operations, 1136.53 ops/s

# after
Workload 0 (concurrency: 10): 1754 operations, 116.93 ops/s
Workload 1 (concurrency: 10): 1352 operations, 90.13 ops/s
Workload 2 (concurrency: 10): 1959 operations, 130.60 ops/s
Workload 3 (concurrency: 100): 999 operations, 66.60 ops/s
Workload 4 (concurrency: 100): 4112 operations, 274.13 ops/s
Workload 5 (concurrency: 50): 10243 operations, 682.87 ops/s
```

#skip-changelog