@Swatinem 2024-03-14
https://github.com/getsentry/symbolicator/pull/1408 is now hooked up to the aggregator and sends Sentry metrics from the final aggregator thread.